4.09 INTEGER fix

+++ app/src/main/java/com/example/retrofit/retrofit/product.kt
@@ -1,15 +1,15 @@
 package com.example.retrofit.retrofit
 
-data class Product (
+data class Product(
     val id: Int,
     val title: String,
     val description: String,
     val price: Int,
     val discountPercentage: Float,
-    val rating: Int,
-    val stock: Float,
+    val rating: Float,
+    val stock: Int,
     val brand: String,
     val category: String,
     val thumbnail: String,
     val images: List<String>
-    )
+)
